+++ cpp/gui/dwi/render_frame.cpp
@@ -347,7 +347,8 @@ void RenderFrame::screenshot(int oversampling, const std::filesystem::path &imag
   screenshot_path = image_path;
   OS = oversampling;
   OS_x = OS_y = 0;
-  framebuffer.reset(new GLubyte[3 * projection.width() * projection.height()]);
+  framebuffer.reset(
+      new GLubyte[3 * static_cast<size_t>(projection.width()) * static_cast<size_t>(projection.height())]);
   pix.reset(new QImage(OS * projection.width(), OS * projection.height(), QImage::Format_RGB32));
   update();
 }
@@ -362,7 +363,8 @@ void RenderFrame::snapshot() {
   for (int j = 0; j < projection.height(); j++) {
     int j2 = projection.height() - j - 1;
     for (int i = 0; i < projection.width(); i++) {
-      GLubyte *p = framebuffer.get() + 3 * (i + projection.width() * j);
+      GLubyte *p = framebuffer.get() +
+                   3 * (static_cast<size_t>(i) + static_cast<size_t>(projection.width()) * static_cast<size_t>(j));
       pix->setPixel(start_i + i, start_j + j2, qRgb(p[0], p[1], p[2]));
     }
   }
